Add files via upload
This is a modified version of Bravura in which the symbols for the Extended Helmholtz-Ellis Notation have been corrected by the author of the notation (Marc Sabat) to position correctly in relation to the notehead. The symbols are based on my own open-source design, a font I call HEJI. The microtonal accidentals are also located (doubled) in the ascii set which Bravura deliberately leaves open. I suggest keeping this because of the unique character of these accidentals, namely that they may be entered as strings. To release this modification of Bravura / HEJI hybrid is permitted under the SIL license, but the font would have to be renamed. The current version retains the Bravura name to simplify coding and avoid renaming all references while working to integrate the microtonal accidental support.